% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute.variants.error.rates}
\alias{compute.variants.error.rates}
\title{compute.variants.error.rates}
\usage{
compute.variants.error.rates(D, inference)
}
\arguments{
\item{D}{Mutation data from multiple experiments for a list of driver genes provided as a data matrix per time point.}

\item{inference}{Results of the inference by LACE.}
}
\value{
A matrix variants_error_rates with the estimated error rates for the considered variants.
}
\description{
Compute error rates for the considered variants comparing observed data to LACE corrected genotype.
}
\examples{
data(longitudinal_sc_variants)
data(inference)
variants_error_rates <- compute.variants.error.rates(longitudinal_sc_variants,inference)

}
