% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_HN120Metastasis}
\alias{data_HN120Metastasis}
\title{mutation data for cell line HN120 Metastasis from Sharma, A. et al. (2018).}
\format{list of mutation data for 3 time points}
\source{
Sharma, A. et al. Longitudinal single-cell RNA sequencing of patient-derived primary cells reveals drug-induced infidelity in stem cell hierarchy. Nature communications 9, 4931 (2018).
}
\usage{
data(data_HN120Metastasis)
}
\value{
list of mutational data for cell line HN120 Metastasis
}
\description{
mutation data for cell line HN120 Metastasis from Sharma, A. et al. (2018).
}
