% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inference_HN120Metastasis}
\alias{inference_HN120Metastasis}
\title{results obtained with the function LACE on input data data_HN120Metastasis from Sharma, A. et al. (2018).}
\format{results obtained with the function LACE on input data data_HN120Metastasis}
\usage{
data(inference_HN120Metastasis)
}
\value{
results obtained with the function LACE on input data data_HN120Metastasis
}
\description{
results obtained with the function LACE on input data data_HN120Metastasis from Sharma, A. et al. (2018).
}
