% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inference_HN120Primary}
\alias{inference_HN120Primary}
\title{results obtained with the function LACE on input data data_HN120Primary from Sharma, A. et al. (2018).}
\format{results obtained with the function LACE on input data data_HN120Primary}
\usage{
data(inference_HN120Primary)
}
\value{
results obtained with the function LACE on input data data_HN120Primary
}
\description{
results obtained with the function LACE on input data data_HN120Primary from Sharma, A. et al. (2018).
}
