% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LACEinterface.R
\name{LACE2}
\alias{LACE2}
\title{LACE Preprocessing and Analysis Interface}
\usage{
LACE2()
}
\description{
\code{LACE2} displays a RStudio Shiny user interface to
handle the VCF and BAM files processing that is needed to
construct the input for the LACE inference algorithms.
The function generates also the most likelihood longitudinal
clonal tree, and shows the output for further explorations of
the results.
}
\note{
The function \code{LACE} is still available for retrocompatibility.
}
\section{Installation}{
 To install the package with all the
dependecies, devtools and Bioconductor need to be present.

The R version required is greater than 4.1.0.

To install directly from github run:\preformatted{remotes::install_github("https://github.com/BIMIB-DISCo/LACE2",  dependencies = T)
}
}

\section{Dependencies}{

The genetic variant annotation software ANNOVAR
(https://annovar.openbioinformatics.org/en/latest/) and the suite
of programs for interacting with high-throughput sequencing data
Samtools (http://www.htslib.org/) need to be installed.
}

