% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LACEinterface.R
\name{LACEView}
\alias{LACEView}
\title{LACE Preprocessing and Analysis Interface}
\usage{
LACEView()
}
\description{
\code{LACEView} displays a RStudio Shiny user interface to
handle the VCF and BAM files preprocessing that is needed to
construct the input for the LACE inference algorithms.
}
\section{Installation}{
 To install the package with all the
dependecies, devtools and Bioconductor need to be present.

The R version required is greater than 3.6.0.

To install directly from github run:\preformatted{remotes::install_github("https://github.com/BIMIB-DISCo/LACE",  dependencies = T)
}

If necessary you can add\preformatted{auth_token="BLAHBLAHGiuntaAlConfinDelCieloDietroApennino"
}

to the previous call; a github \code{auth_token} can be generated from:

https://github.com/settings/tokens
}

\section{Dependencies}{

The genetic variant annotation software ANNOVAR
(https://annovar.openbioinformatics.org/en/latest/) and the suite
of programs for interacting with high-throughput sequencing data
Samtools (http://www.htslib.org/) need to be installed.
}

