% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute.mutation.distance}
\alias{compute.mutation.distance}
\title{compute.mutation.distance}
\usage{
compute.mutation.distance(inference)
}
\arguments{
\item{inference}{Results of the inference by LACE.}
}
\description{
Compute mutation distance from LACE corrected genotype.
}
\examples{
data(inference)
mutation_distance <- compute.mutation.distance(inference)

}
