% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute.mutation.distance}
\alias{compute.mutation.distance}
\title{compute.mutation.distance}
\usage{
compute.mutation.distance(inference)
}
\arguments{
\item{inference}{Results of the inference by LACE.}
}
\value{
A matrix mutation_distance with the mutation distance among variants computed from LACE corrected genotype and related hierarchical clustering.
}
\description{
Compute mutation distance among variants from LACE corrected genotype and use it to perform hierarchical clustering.
}
\examples{
data(inference)
mutation_distance <- compute.mutation.distance(inference)

}
