% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inference}
\alias{inference}
\title{Results obtained with the function LACE on the provided input data from Rambow, Florian, et al. "Toward minimal residual disease-directed therapy in melanoma." Cell 174.4 (2018): 843-855.}
\format{
Results obtained with the function LACE on the provided input data
}
\usage{
data(inference)
}
\value{
Results obtained with the function LACE on the provided input data
}
\description{
Results obtained with the function LACE on the provided input data from Rambow, Florian, et al. "Toward minimal residual disease-directed therapy in melanoma." Cell 174.4 (2018): 843-855.
}
